[MCA] INCOMPLETE. Added the UnsupportedInstructionView class.
This commit needs to have some test cases to go along with it. Also, make sure
that the unsupported instr call to IPP->modifyInstrDesc() is appropriately
printed to debug.

The purpose of the class is to make it obvious to the user of mca when at least
one of the instructions within the input file is unsupported.

Normally, when mca encounters an unsupported instruction (these are instructions
which do not have a valid scheduling class), it terminates with an error.
However, fixing these instructions in the scheduling model can be quite
disruptive so instead, I've implemented this unsupported instruction handling.

The way the handling works is that when mca finds an instruction that is
unsupported, rather than terminating with an error, it converts the instruction
into one with 1 latency, 1 uOp, 0 resource units, and 0 register dependencies.
We also issue a warning when this instruction is parsed, but the warnings
are easy to miss once the rest of mca's output shows up. So we use this new
UnsupportedInstructionView class to display all of the unsupported instructions
at the end of mca's output. This View also explains how these instructions were
modified (if any were unsupported and modified).

The unsupported instruction handling is optional and is enabled with the
--allow-unsupported-instructions command line argument.

This commit also changed the DisableCustomBehaviour (--disable-cb) command
line argument to be part of the ToolOptions category rather than ViewOptions.